feat: implement API services for authentication and place search functionality
