+++ roamance/src/api/roamance-api.ts
@@ -1,13 +1,5 @@
-/**
- * API Service
- * Handles all API requests for the application
- */
+const API_URL = process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:3001';
 
-const API_URL = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3001';
-
-/**
- * Fetch wrapper with standard options
- */
 export async function fetchAPI(endpoint: string, options: RequestInit = {}) {
   const defaultOptions: RequestInit = {
     headers: {
